Update tests for the new solar azimuth limits.
